# hr-ai-web/contracts/hr-business-service.openapi.yaml
openapi: 3.0.3
info:
  title: hr-business-service API
  version: 1.3.0
  description: |
    hr-ai-web 对接 hr-business-service 的核心接口契约（M02 + AI 四功能闭环）。
servers:
  - url: /api
paths:
...
  /v1/hr/employment-events/{id}/submit:
    post:
      summary: 提交任职事件审批
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: success
  /v1/hr/employment-events/{id}/approve:
    post:
      summary: 审批任职事件（通过/驳回）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - approved
              properties:
                approved:
                  type: boolean
                rejectReason:
                  type: string
      responses:
        '200':
          description: success
  /v1/hr/employment-events/{id}/cancel:
    post:
      summary: 取消任职事件
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: success
...